moved TimeSeriesBuilder to own class
