+++ ta4j-core/src/test/java/org/ta4j/core/TestUtilsTest.java
@@ -69,7 +69,7 @@ public class TestUtilsTest extends AbstractIndicatorTest<TimeSeries, Num> {
     }
 
     private TimeSeries randomSeries() {
-        BaseTimeSeries.SeriesBuilder builder = new BaseTimeSeries.SeriesBuilder();
+        BaseTimeSeriesBuilder builder = new BaseTimeSeriesBuilder();
         TimeSeries series = builder.withNumTypeOf(numFunction).build();
         ZonedDateTime time = ZonedDateTime.of(1970, 1, 1, 1, 1, 1, 1, ZoneId.systemDefault());
         double random;
