+++ ta4j-core/src/test/java/org/ta4j/core/indicators/ATRIndicatorTest.java
@@ -45,7 +45,7 @@ public class ATRIndicatorTest extends AbstractIndicatorTest<TimeSeries, Num> {
 
     @Test
     public void testDummy() throws Exception {
-        TimeSeries series = new BaseTimeSeries.SeriesBuilder().withNumTypeOf(numFunction).build();
+        TimeSeries series = new BaseTimeSeriesBuilder().withNumTypeOf(numFunction).build();
         series.addBar(new MockBar(ZonedDateTime.now().minusSeconds(5), 0, 12, 15, 8, 0, 0, 0,numFunction));
         series.addBar(new MockBar(ZonedDateTime.now().minusSeconds(4), 0, 8, 11, 6, 0, 0, 0,numFunction));
         series.addBar(new MockBar(ZonedDateTime.now().minusSeconds(3), 0, 15, 17, 14, 0, 0, 0,numFunction));
