+++ ta4j-core/src/test/java/org/ta4j/core/indicators/FisherIndicatorTest.java
@@ -25,7 +25,7 @@ package org.ta4j.core.indicators;
 
 import org.junit.Before;
 import org.junit.Test;
-import org.ta4j.core.BaseTimeSeries;
+import org.ta4j.core.BaseTimeSeriesBuilder;
 import org.ta4j.core.Indicator;
 import org.ta4j.core.TimeSeries;
 import org.ta4j.core.mocks.MockBar;
@@ -48,7 +48,7 @@ protected TimeSeries series;
     @Before
     public void setUp() {
 
-        series = new BaseTimeSeries.SeriesBuilder()
+        series = new BaseTimeSeriesBuilder()
                 .withNumTypeOf(numFunction)
                 .withName("NaN test")
                 .build();
