+++ ta4j-core/src/test/java/org/ta4j/core/indicators/helpers/DecimalTransformIndicatorTest.java
@@ -25,7 +25,7 @@ package org.ta4j.core.indicators.helpers;
 
 import org.junit.Before;
 import org.junit.Test;
-import org.ta4j.core.BaseTimeSeries;
+import org.ta4j.core.BaseTimeSeriesBuilder;
 import org.ta4j.core.Indicator;
 import org.ta4j.core.TimeSeries;
 import org.ta4j.core.indicators.AbstractIndicatorTest;
@@ -56,7 +56,7 @@ public class DecimalTransformIndicatorTest extends AbstractIndicatorTest<Indicat
 
     @Before
     public void setUp() {
-        TimeSeries series = new BaseTimeSeries.SeriesBuilder().withNumTypeOf(numFunction).build();
+        TimeSeries series = new BaseTimeSeriesBuilder().withNumTypeOf(numFunction).build();
         ConstantIndicator<Num> constantIndicator = new ConstantIndicator<>(series, numOf(4));
 
         transPlus = new DecimalTransformIndicator(constantIndicator, 10, DecimalTransformType.plus);
